Silence a compiler warning.

+++ Src/builtin.c
@@ -4970,7 +4970,7 @@ bin_read(char *name, char **args, Options ops, UNUSED(int func))
 			    }
 			    else if (ret == 0) /* handle null as normal char */
 				ret = 1;
-			    else if (ret > val) {
+			    else if (ret > (size_t)val) {
 				/* Some mbrlen()s return the full char len */
 				ret = val;
 			    }
